+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+### Added
+- Available LLMs can now be configured in the backend.
+
 ## [25.5.0] - 2025-05-05
 
 ### Added